fix: shake out more bugs
